Move CacheMapUtil classes to different module
Prevents Interface to be picked up by mybatis mapper scan
