+++ persistence/persistence-api/src/main/java/org/cbioportal/persistence/cachemaputil/CacheMapBuilder.java
@@ -1,4 +1,4 @@
-package org.cbioportal.persistence.mybatis.util;
+package org.cbioportal.persistence.cachemaputil;
 
 import java.util.function.Function;
 import java.util.stream.Collectors;
