+++ persistence/persistence-api/src/main/java/org/cbioportal/persistence/cachemaputil/CacheMapUtil.java
@@ -1,4 +1,4 @@
-package org.cbioportal.persistence.mybatis.util;
+package org.cbioportal.persistence.cachemaputil;
 
 import org.cbioportal.model.CancerStudy;
 import org.cbioportal.model.MolecularProfile;
