+++ persistence/persistence-api/src/main/java/org/cbioportal/persistence/cachemaputil/StaticRefCacheMapUtil.java
@@ -30,17 +30,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.cbioportal.persistence.mybatis.util;
+package org.cbioportal.persistence.cachemaputil;
 
 import org.cbioportal.model.CancerStudy;
 import org.cbioportal.model.MolecularProfile;
 import org.cbioportal.model.SampleList;
 import org.cbioportal.utils.config.annotation.ConditionalOnExpression;
-import org.cbioportal.utils.security.PortalSecurityConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
@@ -100,8 +98,7 @@ public class StaticRefCacheMapUtil implements CacheMapUtil {
     public Map<String, String> getGenericAssayStableIdToMolecularProfileIdMap() {
         return genericAssayStableIdToMolecularProfileIdCache;
     }
-
-    //  bean is only instantiated when there is user authorization
+    
     @Override
     public boolean hasCacheEnabled() {
         return true;
