+++ security/security-spring/src/main/java/org/cbioportal/security/spring/CancerStudyPermissionEvaluator.java
@@ -36,7 +36,7 @@ import java.io.Serializable;
 import java.util.*;
 import org.apache.commons.logging.*;
 import org.cbioportal.model.*;
-import org.cbioportal.persistence.mybatis.util.CacheMapUtil;
+import org.cbioportal.persistence.cachemaputil.CacheMapUtil;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.security.access.PermissionEvaluator;
 import org.springframework.security.core.Authentication;
